nca-toolkit/requirements.txt
Flask
python-dotenv
minio
requests
flask-cors
pandas
pytest